Moved i/ostream to forward declarations
o/istream classes are only used by reference in the Vector3D header and therefore should be forward declared to reduce header complexity. This is accomplished by including the <iosfwd> header which correctly forward declares any ios classes, it is not considered correct to attempt these forward declarations yourself. Because iostream is no longer present, <initializer_list> is manually included in the header